MVPainter/requirements.txt
# datasets==2.19.0
# diffusers==0.20.2
# einops==0.8.0
# huggingface-hub==0.22.2
# imageio==2.34.1
# jaxtyping==0.2.28
# joblib==1.4.0
# lightning-utilities==0.11.2
# matplotlib==3.8.4
# numpy
# omegaconf==2.3.0
# opencv-python
# pandas==2.2.2
# pillow==10.3.0
# plyfile==1.0.3
# PyMCubes==0.1.4
# pyparsing==3.1.2
# pytorch-lightning==2.2.4
# PyYAML==6.0.1
# safetensors==0.4.3
# scikit-image==0.23.2
# scipy==1.13.0
# tensorboard==2.16.2
# tensorboardX==2.6.2.2
# timm==0.9.16
# tokenizers==0.19.1
# tqdm==4.66.2
# transformers==4.40.1
# trimesh==4.5.2
# wandb==0.16.6
# segment_anything==1.0
# spaces==0.28.3
# open_clip_torch==2.7.0
# kiui==0.2.7
# accelerate==0.29.1
# addict==2.4.0
# xformers
# webdataset
# IPython
# spconv==2.3.6
# OpenEXR==3.2.4
# rembg
# onnxruntime
# bpy==3.6.0
# xatlas
# pygltflib



datasets
diffusers==0.20.2
einops
huggingface-hub==0.22.2
imageio
jaxtyping
# joblib
lightning-utilities
# matplotlib
numpy
omegaconf
opencv-python
pandas
pillow
# plyfile
# PyMCubes
pyparsing
pytorch-lightning==2.2.4
PyYAML
safetensors
scikit-image
scipy
tensorboard
tensorboardX
timm
tokenizers
tqdm
transformers
trimesh
wandb
# segment_anything==1.0
spaces
# open_clip_torch
kiui
accelerate
addict
xformers
webdataset
IPython
spconv
OpenEXR==3.2.4
rembg
onnxruntime
bpy==3.6.0
xatlas
pygltflib